deployment scripts/requirements.txt
discord.py==1.7.3
nltk==3.6.1
numpy==1.20.1
pandas==1.2.4
tensorflow-cpu==2.7.0
